+++ client/src/features/advancedSearch/RelationshipRow.tsx
@@ -86,7 +86,8 @@ interface IRelationshipRow {
   parentScope: string
   parentStateId: string
   nestedLevel: number
-  bgColor: 'bg-light' | 'bg-white'
+  parentBgColor: 'bg-light' | 'bg-white'
+  rowBgColor: string
   hasYoungerSiblings?: boolean
   parentGroupName?: string
 }
@@ -99,7 +100,8 @@ interface IRelationshipRow {
  * @param {string} parentScope the scope of the parent object
  * @param {string} parentStateId id of the parent object within the advanced search state
  * @param {number} nestedLevel level of depth within the advanced search state
- * @param {string} bgColor the background color of the child container
+ * @param {string} parentBgColor the background color of the parent container
+ * @param {string} rowBgColor the background color of the current row
  * @param {boolean} hasYoungerSiblings the background color of the child container
  * @returns {JSX.Element}
  */
@@ -110,7 +112,8 @@ const RelationshipRow: React.FC<IRelationshipRow> = ({
   parentScope,
   parentStateId,
   nestedLevel,
-  bgColor,
+  parentBgColor,
+  rowBgColor,
   hasYoungerSiblings = false,
   parentGroupName = undefined,
 }) => {
@@ -135,7 +138,8 @@ const RelationshipRow: React.FC<IRelationshipRow> = ({
         className={`col-12 ${!hasChildInputField ? 'mb-3' : ''}`}
       >
         <StyledInputGroupDiv
-          className="jusify-content-between flex-nowrap bg-white"
+          className="jusify-content-between flex-nowrap"
+          backgroundColor={rowBgColor}
           data-testid={`${selectedKey}-${stateId}-relationship-row`}
         >
           <span className="w-100 d-flex ps-2">
@@ -188,7 +192,7 @@ const RelationshipRow: React.FC<IRelationshipRow> = ({
               parentScope={scopeToPassToNestedForm}
               parentStateId={stateId}
               nestedLevel={nestedLevel + 1}
-              parentBgColor={bgColor}
+              parentBgColor={parentBgColor}
             />
           </Col>
           {/* <StyledBorderDiv /> */}