+++ client/src/features/results/ConceptResults.tsx
@@ -31,6 +31,9 @@ const ConceptResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
   const pageParam = `${paramPrefix}p`
   const page: any = queryString.has(pageParam) ? queryString.get(pageParam) : 1
   const sort = queryString.get(`${tab}Sort`)
+  const view: string = queryString.has('view')
+    ? (queryString.get('view') as string)
+    : 'list'
 
   const { data, isFetching, isSuccess, isError, error, isLoading, status } =
     searchResponse
@@ -40,11 +43,6 @@ const ConceptResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
   if (isError) {
     errorMessage = error.data.errorMessage
   }
-
-  const resultsList = (
-    results: Array<IOrderedItems>,
-  ): Array<React.ReactElement<any>> =>
-    results.map((result) => <ConceptSnippet key={result.id} uri={result.id} />)
 
   let estimate = 0
   if (isSuccess && data) {
@@ -56,6 +54,19 @@ const ConceptResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
     return null
   }
 
+  const resultsList = (
+    results: Array<IOrderedItems>,
+  ): Array<React.ReactElement<any>> =>
+    results.map((result, ind) => (
+      <ConceptSnippet
+        key={result.id}
+        uri={result.id}
+        view={view}
+        totalResults={estimate}
+        index={ind + 1}
+      />
+    ))
+
   return (
     <StyledEntityResultsRow>
       {(isSuccess || isError) && (
@@ -66,7 +77,6 @@ const ConceptResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
             label="Concepts"
             overlay="conceptsAndGroupings"
             resultsData={data}
-            toggleView
           />
         </Col>
       )}
@@ -86,7 +96,12 @@ const ConceptResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
           <Col xs={12} sm={12} md={9} lg={9}>
             {!isFetching && isSuccess && data && (
               <React.Fragment>
-                {resultsList(data.orderedItems)}
+                {view === 'list' && resultsList(data.orderedItems)}
+                {view === 'grid' && (
+                  <Row xs={1} sm={2} md={3} lg={4} className="g-4 mx-3 pt-2">
+                    {resultsList(data.orderedItems)}
+                  </Row>
+                )}
                 {estimate >= 20 && (
                   <Paginate
                     estimate={estimate}