+++ client/src/features/results/ConceptSnippet.tsx
@@ -1,5 +1,5 @@
-import React from 'react'
-import { Col, Row } from 'react-bootstrap'
+import React, { useState } from 'react'
+import { Card, Col, Row } from 'react-bootstrap'
 
 import config from '../../config/config'
 import ConceptParser from '../../lib/parse/data/ConceptParser'
@@ -11,15 +11,36 @@ import { stripYaleIdPrefix } from '../../lib/parse/data/helper'
 import { useGetItemQuery } from '../../redux/api/ml_api'
 import GenericBreadcrumbHierarchy from '../common/GenericBreadcrumbHierarchy'
 import { getNextConceptUris } from '../../lib/util/hierarchyHelpers'
+import PreviewImageOrIcon from '../common/PreviewImageOrIcon'
+import StyledSnippetTitle from '../../styles/features/results/SnippetTitle'
+import RecordLink from '../common/RecordLink'
+import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
+import theme from '../../styles/theme'
 
 import SnippetHeader from './SnippetHeader'
 
 interface IProps {
   uri: string
+  view: string
+  totalResults?: number
+  index?: number
+  pageLength?: number
   titleOfTabbedContent?: string
 }
 
-const ConceptSnippet: React.FC<IProps> = ({ uri, titleOfTabbedContent }) => {
+const ConceptSnippet: React.FC<IProps> = ({
+  uri,
+  view,
+  totalResults,
+  index,
+  pageLength = 20,
+  titleOfTabbedContent,
+}) => {
+  const [isMobile, setIsMobile] = useState<boolean>(
+    window.innerWidth < theme.breakpoints.md,
+  )
+  useResizeableWindow(setIsMobile)
+
   const { data, isSuccess, isLoading } = useGetItemQuery({
     uri: stripYaleIdPrefix(uri),
     profile: 'results',
@@ -27,6 +48,7 @@ const ConceptSnippet: React.FC<IProps> = ({ uri, titleOfTabbedContent }) => {
 
   if (isSuccess && data) {
     const concept = new ConceptParser(data)
+    const images = concept.getImages()
     const descriptions = concept.getDescriptions(config.aat.langen)
 
     const snippetDataComponent = (
@@ -38,7 +60,7 @@ const ConceptSnippet: React.FC<IProps> = ({ uri, titleOfTabbedContent }) => {
                 <StyledDt>Descriptions</StyledDt>
                 <StyledDd data-testid="concept-snippet-description">{`${descriptions[0].slice(
                   0,
-                  200,
+                  view === 'list' ? 200 : 100,
                 )}...`}</StyledDd>
               </Col>
             </Row>
@@ -54,18 +76,70 @@ const ConceptSnippet: React.FC<IProps> = ({ uri, titleOfTabbedContent }) => {
       </React.Fragment>
     )
 
-    return (
-      <React.Fragment>
-        <div className="m-2 d-flex">
-          <SnippetHeader
-            data={data}
-            snippetData={snippetDataComponent}
-            titleOfTabbedContent={titleOfTabbedContent}
+    if (view === 'list') {
+      return (
+        <React.Fragment>
+          <div className="m-2 d-flex">
+            <SnippetHeader
+              data={data}
+              snippetData={snippetDataComponent}
+              titleOfTabbedContent={titleOfTabbedContent}
+            />
+          </div>
+          <StyledHr
+            width="100%"
+            className={`conceptSnippetHr ${
+              isMobile &&
+              totalResults &&
+              index &&
+              totalResults <= pageLength &&
+              index === totalResults
+                ? 'lastResult'
+                : ''
+            }`}
           />
-        </div>
-        <StyledHr width="100%" className="conceptSnippetHr" />
-      </React.Fragment>
-    )
+        </React.Fragment>
+      )
+    }
+
+    if (view === 'grid') {
+      return (
+        <Col className="h-auto">
+          <Card className="h-100" data-testid="grid-view">
+            {images.length > 0 ? (
+              <PreviewImageOrIcon
+                images={images}
+                entity={data}
+                className="card-img-top py-0"
+                width="auto"
+                height="auto"
+              />
+            ) : (
+              <PreviewImageOrIcon
+                images={images}
+                entity={data}
+                className="card-img-top"
+                height="152px"
+                width="auto"
+              />
+            )}
+            <Card.Body>
+              <StyledSnippetTitle
+                className="card-title d-flex"
+                data-testid="grid-view-object-results-snippet-title"
+              >
+                <RecordLink
+                  url={data.id}
+                  className="overflow-auto"
+                  linkCategory="Results Snippet"
+                />
+              </StyledSnippetTitle>
+              <Card.Text>{snippetDataComponent}</Card.Text>
+            </Card.Body>
+          </Card>
+        </Col>
+      )
+    }
   }
 
   if (isLoading) {