+++ client/src/features/results/EventResults.tsx
@@ -32,6 +32,9 @@ const EventResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
   const page: any = queryString.has(pageParam) ? queryString.get(pageParam) : 1
   const sort = queryString.get(`${tab}Sort`)
   const hasSimpleSearchQuery = queryString.has('sq')
+  const view: string = queryString.has('view')
+    ? (queryString.get('view') as string)
+    : 'list'
 
   const { data, isFetching, isSuccess, isError, error, isLoading, status } =
     searchResponse
@@ -42,11 +45,6 @@ const EventResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
     errorMessage = error.data.errorMessage
   }
 
-  const resultsList = (
-    results: Array<IOrderedItems>,
-  ): Array<React.ReactElement<any>> =>
-    results.map((result) => <EventSnippet key={result.id} uri={result.id} />)
-
   let estimate = 0
   if (isSuccess && data) {
     estimate = getEstimates(data)
@@ -56,6 +54,19 @@ const EventResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
   if (status === 'uninitialized') {
     return null
   }
+
+  const resultsList = (
+    results: Array<IOrderedItems>,
+  ): Array<React.ReactElement<any>> =>
+    results.map((result, ind) => (
+      <EventSnippet
+        key={result.id}
+        uri={result.id}
+        view={view}
+        totalResults={estimate}
+        index={ind + 1}
+      />
+    ))
 
   return (
     <StyledEntityResultsRow
@@ -88,7 +99,12 @@ const EventResults: React.FC<IProps> = ({ searchResponse, isMobile }) => {
           <Col xs={12} sm={12} md={9} lg={9}>
             {!isFetching && isSuccess && data && (
               <React.Fragment>
-                {resultsList(data.orderedItems)}
+                {view === 'list' && resultsList(data.orderedItems)}
+                {view === 'grid' && (
+                  <Row xs={1} sm={2} md={3} lg={4} className="g-4 mx-3 pt-2">
+                    {resultsList(data.orderedItems)}
+                  </Row>
+                )}
                 {estimate >= 20 && (
                   <Paginate
                     estimate={estimate}