+++ client/src/features/results/MobileNavigation.tsx
@@ -146,6 +146,7 @@ const MobileNavigation: React.FC<IProps> = ({
 
   useResizeableWindow(handleModalStateOnWindowResize)
 
+  const currentTabIcon = getIcon(searchScope[tab])
   return (
     <React.Fragment>
       <MobileTabButton
@@ -155,12 +156,17 @@ const MobileNavigation: React.FC<IProps> = ({
         }}
         role="button"
         estimate={estimates[tab]}
-        icon={getIcon(advancedSearchTitles[tab])}
+        icon={currentTabIcon}
         tab={tab}
         handleClick={handleShow}
         showArrow
       />
-      <Modal show={show} animation={false} onHide={() => setShow(false)}>
+      <Modal
+        show={show}
+        animation={false}
+        onHide={() => setShow(false)}
+        className="d-flex align-items-center mobile-navigation-modal"
+      >
         <Modal.Body>
           {Object.entries(searchScope).map(([key, value]) => (
             <MobileTabButton
@@ -170,7 +176,7 @@ const MobileNavigation: React.FC<IProps> = ({
                 isFetching,
               }}
               estimate={estimates[key]}
-              icon={value}
+              icon={getIcon(value)}
               tab={key}
               handleClick={() => handleOnLinkSelect(key)}
               isCurrentTab={tab === key}